+++ src/GlobedPlayerCountNode.hpp
@@ -13,9 +13,11 @@ private:
     cocos2d::CCLabelBMFont* m_label = nullptr;
     cocos2d::CCSprite* m_icon = nullptr;
     size_t m_displayedCount = std::numeric_limits<size_t>::max();
+    float m_rainbowHue = 0.f;
 
     bool init();
     void refresh(float);
+    void updateRainbow(float);
     void ensureIcon();
 };
 }